requirements.txt
asgiref==3.7.2
Django==4.2.5
django-ckeditor==6.7.0
django-credit-cards==0.4.1
django-js-asset==2.1.0
Faker==19.6.2
faker-commerce==1.0.3
neural-admin==1.1.2
psycopg2==2.9.7
pycodestyle==2.11.0
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
Unidecode==1.3.6
